feat: Added UsuarioController and both of Controllers are working

+++ GerenciadorDeTarefas/Migrations/20241225214004_InitialDB.cs
@@ -31,7 +31,7 @@ namespace GerenciadorDeTarefas.Migrations
                 name: "Usuarios",
                 columns: table => new
                 {
-                    Id = table.Column<int>(type: "int", nullable: false)
+                    Id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(255)", maxLength: 255, nullable: false),
                     Email = table.Column<string>(type: "nvarchar(150)", maxLength: 150, nullable: false)
